+++ setup/dialer_process/dialer/eccp-examples/getcampaigninfo.php
@@ -1,0 +1,26 @@
+#!/usr/bin/php
+<?php
+require_once ("/var/www/html/modules/agent_console/libs/ECCP.class.php");
+
+if (count($argv) < 3) {
+    fprintf(STDERR, $argv[0]." [incoming|outgoing] [campaign-id]\n");
+    exit(0);
+}
+
+$x = new ECCP();
+try {
+    print "Connect...\n";
+    $cr = $x->connect("localhost", "agentconsole", "agentconsole");
+    if (isset($cr->failure)) die('Failed to connect to ECCP - '.$cr->failure->message."\n");
+
+    print "Getting campaign info...\n";
+    $r = $x->getcampaigninfo($argv[1], $argv[2]);
+    print_r($r);
+
+    print "Disconnect...\n";
+    $x->disconnect();
+} catch (Exception $e) {
+    print_r($e);
+    print_r($x->getParseError());
+}
+?>